+++ rocc/core/src/main/java/edu/chl/rocc/core/factories/RoCCFactory.java
@@ -24,7 +24,7 @@ public class RoCCFactory implements IRoCCFactory {
     }
 
     @Override
-    public ICharacter createCharacter(String name, int x, int y) {
+    public ICharacter createCharacter(String name, float x, float y) {
         return new Character(name);
     }
